major refactor of jsc3d
