+++ frontend/web/src/app/app.component.html
@@ -4,4 +4,6 @@
     Welcome to {{title}}!
   </h1>
 
+  <app-stl-viewer></app-stl-viewer>
+
 </div>
