+++ frontend/web/src/assets/lib/jsc3d/classes/Matrix.ts
@@ -1,0 +1,10 @@
+    /**
+        @class Matrix3x4
+
+        This class implements 3x4 matrix and mass operations for 3D transformations.
+    */
+    export class Matrix3x4 {
+        m00 = 1; m01 = 0; m02 = 0; m03 = 0;
+        m10 = 0; m11 = 1; m12 = 0; m13 = 0;
+        m20 = 0; m21 = 0; m22 = 1; m23 = 0;
+    }
